requirements.txt
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.2.1
filelock==3.8.0
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
platformdirs==2.5.3
protobuf==3.20.1
pymongo==4.3.3
python-decouple==3.6
six==1.16.0
virtualenv==20.16.6
Werkzeug==2.2.2
